extract GraphView functionality in separate class to make it re-usable in ARCitect/elsewhere
